StaveTie -- move rendering logic out of draw
Moving the rendering logic (where to start, where to end, etc.) out of the draw() method for StaveTie so that it can be tested separately from the visual tests.  (And hopefully expanded soon).

Added ability to configure the cutoff point for where a tie gets classified as a short tie and what the effects of that are.

Added a visual test (w/ some application logic tests) of a more standard use of ties: two chords that have the same pitches, and with two different ties between the same chords representing different directions.  (Surprisingly, this most standard use of ties was not being tested).

Still learning the factory system.  :-)