Remove unneeded test on StaveModifier
A test that each StaveModifier has a draw() method goes back to the pre-ES6 class version of Vexflow and is unnecessary now.

https://github.com/vexflow/vexflow/commit/f5e562519dffd0e5931c136218ab36bbc2cf1cb3

Added a type to remind ourselves that these are StaveModifiers not Modifiers.

Q: Stave().getModifiers returns a different array than Stave.modifiers if position or category is specified, but returns the original array if neither is specified.  I wonder if line 567 should be changed to return [...this.modifiers] so that in all cases `.modifiers` remains immutable?