Remove hash from private properties
For many of them I looked up in the past whether they were protected or private (or if they were mentioned in docs) and set protected or private accordingly.  For the rest I used my best judgment about whether I would ever want to query them outside of the system or not.